Random generator mp

+++ app/src/main/java/com/example/fallingBall/multiPlayer/Obstacle.java
@@ -16,7 +16,7 @@ public class Obstacle implements GameObject {
         return rectangle2;
     }
 
-    public void incrementY (float y){
+    public void incrementY (double y){
         rectangle.top += y;
         rectangle.bottom += y;
         rectangle2.top += y;
